Move image's call_action, get_value and macro to mixin class which can be shared by all child objects with their own base paths in the frontend store tree
